Move components to a dedicated folder
