+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "./components/App";
 import { Provider } from 'react-redux';
 import { store } from "./redux";
 import reportWebVitals from "./reportWebVitals";
